EFK-FE-24 fix: change word cards
