+++ efk-fe/src/components/Words/WordAdminCard/WordAddCard/WordAddCard.module.scss
@@ -1,4 +1,4 @@
-@import '../../../styles/styles-config.scss';
+@import '../../../../styles/styles-config.scss';
 
 .card {
   @include card($height: 300px);
@@ -34,49 +34,14 @@
     transform: rotateY(180deg);
   }
 
-  &__link {
-    @include button();
-  }
-
-  &__btnAudio {
-    margin: 5px;
-    width: 20px;
-    height: 20px;
+  &__frontImg {
+    width: 250px;
+    height: 250px;
+    background-image: url(../../../../../public/icons/icon__cardAdd.png);
     background-repeat: no-repeat;
     background-size: contain;
-    background-image: url(/icons/icon__play.png);
-
-    &:hover {
-      transform: scale(1.2);
-    }
+    background-position: center;
   }
-}
-
-.info {
-  @include flex($direction: column, $align: flex-start);
-
-  &__title {
-    margin-right: 5px;
-    font-weight: 700;
-  }
-
-  &__text {
-    font-family: sans-serif;
-  }
-}
-
-.fileInputs {
-  @include flex($justify: space-between);
-}
-
-.btnSubmit,
-.btnUpdate {
-  @include button($background: $blue_color);
-}
-
-.btnCancel,
-.btnDelete {
-  @include button($background: $red_color);
 }
 
 .btnAdd {
