+++ efk-fe/src/components/Words/WordAdminCard/WordAddCard/index.ts
@@ -1,0 +1,1 @@
+export { WordAddCard } from './WordAddCard';
