+++ efk-fe/src/components/Words/WordAdminCard/WordEditCard/WordEditCard.module.scss
@@ -1,0 +1,79 @@
+@import '../../../../styles/styles-config.scss';
+
+.card {
+  @include card($height: 300px);
+
+  &.flipped {
+    transform: rotateY(180deg);
+
+    .btn {
+      z-index: 1;
+    }
+  }
+
+  &__front,
+  &__back {
+    position: absolute;
+    left: 0;
+    top: 0;
+    padding: 3px;
+    width: 100%;
+    height: 100%;
+
+    @include flex($direction: column, $wrap: nowrap);
+
+    border-radius: 19px;
+    box-shadow: 1px 3px 4px rgb(0 0 0 / 5%);
+    background: white;
+    transition: all 0.5s ease;
+    backface-visibility: hidden;
+    -webkit-backface-visibility: hidden;
+  }
+
+  &__back {
+    transform: rotateY(180deg);
+  }
+
+  &__btnAudio {
+    margin: 5px;
+    width: 20px;
+    height: 20px;
+    background-repeat: no-repeat;
+    background-size: contain;
+    background-image: url(/icons/icon__play.png);
+
+    &:hover {
+      transform: scale(1.2);
+    }
+  }
+  &__frontImg {
+    width: 150px;
+    height: 150px;
+    background-repeat: no-repeat;
+    background-size: contain;
+    background-position: center;
+  }
+}
+
+.info {
+  @include flex($direction: column, $align: flex-start);
+
+  &__title {
+    margin-right: 5px;
+    font-weight: 700;
+  }
+
+  &__text {
+    font-family: sans-serif;
+  }
+}
+
+.btnSubmit,
+.btnUpdate {
+  @include button($background: $blue_color);
+}
+
+.btnCancel,
+.btnDelete {
+  @include button($background: $red_color);
+}
