+++ efk-fe/src/components/Words/WordAdminCard/WordEditCard/index.ts
@@ -1,0 +1,1 @@
+export { WordEditCard } from './WordEditCard';
