+++ efk-fe/src/components/Words/WordAdminCard/WordForm/index.ts
@@ -1,0 +1,1 @@
+export { WordForm } from './WordForm';
