+++ efk-fe/src/components/Words/WordAdminCard/index.ts
@@ -1,0 +1,2 @@
+export { WordAddCard } from './WordAddCard';
+export { WordEditCard } from './WordEditCard';
