+++ efk-fe/src/components/Words/WordCard/WordCard.module.scss
@@ -37,10 +37,11 @@
     transition: all 0.5s ease;
     backface-visibility: hidden;
     -webkit-backface-visibility: hidden;
-    &:hover {
-      background-color: #f5fffa;
-      box-shadow: 19px -4px 39px -20px #71edb3, -14px 11px 15px -12px #71edb3;
-    }
+  }
+
+  &__front:hover {
+    background-color: #f5fffa;
+    box-shadow: 19px -4px 39px -20px #71edb3, -14px 11px 15px -12px #71edb3;
   }
 
   &__back {
@@ -52,11 +53,18 @@
   &__frontImg,
   &__backImg {
     width: 100%;
-    height: 100%;
     border-radius: 19px;
     object-fit: contain;
     background-clip: content-box;
+    transform: scale(0.83);
     transition: all 0.5s ease;
+  }
+
+  &__backImg {
+    height: 250px;
+    background-position: center;
+    background-size: contain;
+    background-repeat: no-repeat;
   }
 
   &__frontInfo,
@@ -76,6 +84,7 @@
   &__backInfo {
     font-family: inherit;
     justify-content: center;
+
     .card__back-desc {
       font-weight: 800;
     }
@@ -87,14 +96,19 @@
     height: 20px;
     background-repeat: no-repeat;
     background-size: contain;
+
+    &:hover {
+      transition: all 0.5s ease-in-out;
+      transform: scale(1.2);
+    }
   }
 
   &__btnFlip {
-    background-image: url(/icons/icon__reverse.png);
+    background-image: url(/icons/icon__flipp.svg);
   }
 
   &__btnAudio {
-    background-image: url(/icons/icon__audio.png);
+    background-image: url(/icons/icon__audio.svg);
   }
 }
