+++ efk-fe/src/components/Words/WordCard/WordCard.tsx
@@ -7,7 +7,7 @@ import { Word } from '../../../interfaces';
 import { useFlipItem } from '../../../hooks';
 import { statisticsDB } from '../../../lib';
 import { gameService } from '../../../services';
-import { isGameModePlay, isGameModeTrain, playAudio } from '../../../utils';
+import { isGameModePlay, isGameModeTrain, playAudio, shimmer, toBase64 } from '../../../utils';
 import { ANSWER, GAME_MODE } from '../../../enums';
 import styles from './WordCard.module.scss';
 
@@ -58,13 +58,13 @@ export const WordCard: FC<WordCardProps> = memo(({ word, categoryName, gameMode 
       <div className={styles.card__front}>
         <Image
           className={styles.card__frontImg}
+          src={image.url}
           width={250}
           height={250}
           quality={100}
           placeholder="blur"
-          blurDataURL={image.url}
+          blurDataURL={toBase64(shimmer(250, 250))}
           alt={name}
-          src={image.url}
         />
         <div className={frontInfoStyle}>
           <button
@@ -78,7 +78,7 @@ export const WordCard: FC<WordCardProps> = memo(({ word, categoryName, gameMode 
         </div>
       </div>
       <div className={styles.card__back}>
-        <Image className={styles.card__backImg} width={250} height={250} quality={100} src={image.url} />
+        <div className={styles.card__backImg} style={{ backgroundImage: `url(${image.url})` }} />
         <div className={styles.card__backInfo}>
           <p>{translation}</p>
         </div>
