+++ efk-fe/src/components/Words/index.ts
@@ -1,4 +1,4 @@
 export { WordsContainer } from './WordsContainer/WordsContainer';
-export { WordEditCard } from './WordAdminCard/WordEditCard';
-export { WordAddCard } from './WordAdminCard/WordAddCard';
+export { WordEditCard } from './WordAdminCard/WordEditCard/WordEditCard';
+export { WordAddCard } from './WordAdminCard/WordAddCard/WordAddCard';
 export { WordCard } from './WordCard/WordCard';
